requirements.txt
flask==2.3.3
python-dotenv==1.0.0
numpy==1.24.3
sentence-transformers==2.2.2
google-generativeai==0.3.1
langchain==0.1.0
langchain-groq==0.0.5
langchain-chroma==0.0.10
langchain-google-genai==0.0.5
langchain-community==0.0.13
chromadb==0.4.18